requirements.txt
yt-dlp>=2024.0.0
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx[socks]>=0.24.0
